NTF: fix MSVC warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@31603 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/ntf/ntf_generic.cpp
@@ -215,7 +215,7 @@ void OGRNTFDataSource::WorkupGeneric( NTFFileReader * poReader )
                       {
                           poClass->CheckAddAttr( poAttDesc->val_type,
                                                  poAttDesc->finter,
-                                                 strlen(papszValues[iAtt]) );
+                                                 static_cast<int>(strlen(papszValues[iAtt])) );
                       }
 
                       if( CSLFindString( papszFullAttList,
